Refactored cli logs modules
